Chore. Todo 관련 초기데이터 생성
